fixes for visual studio

+++ libraries/linalg/types/inc/arraytypes.hpp
@@ -13,6 +13,6 @@ template <size_t N>
 using IntArray = std::array<int, N>;
 
 template <size_t N>
-using UIntArray = std::array<uint, N>;
+using UIntArray = std::array<size_t, N>;
 
 #endif
